+++ build.gradle
@@ -8,7 +8,11 @@ repositories {
  
 dependencies {
 //    compile 'log4j:log4j:1.2.17'
+    compile group: 'com.google.guava', name: 'guava', version: '19.0'
+    compile group: 'org.springframework', name: 'spring-jdbc', version: '4.3.1.RELEASE'
+    compile group: 'mysql', name: 'mysql-connector-java', version: '5.1.6'
+    compile 'log4j:log4j:1.2.17'
+
     testCompile 'junit:junit:4.11'
     testCompile "org.mockito:mockito-core:1.+"
-    compile group: 'com.google.guava', name: 'guava', version: '19.0'
 }
