--- test/CSVReaderTest.java
@@ -1,22 +1,0 @@
-import org.junit.Before;
-import org.junit.Test;
-
-/**
- * Created by ken-linux2 on 24/07/16.
- */
-public class CSVReaderTest {
-
-//    CSVReader classUnderTest;
-//    final String STRING_TEST="Google,2831944540,186307692,,,,GOOGLE_WOMENS_ACCESSORIES_PLA,WOMENS_ACCESSORIES,GOOGLE,PLA,CENTER_CORE,,WOMENS_ACCESSORIES,BRANDS,BRANDS,,";
-//    @Before
-//    public void setup() {
-//        classUnderTest = new CSVReader();
-//    }
-//
-//    @Test
-//    public void shouldReadFromACSVFile(){
-//    classUnderTest.run();
-//
-//    }
-
-}
