+++ test/data/shouldCheckIfFileIsEmpty.csv
@@ -1,0 +1,1 @@
+channel	channel_account_id	channel_campaign_id	channel_adgroup_id	channel_ad_id	product_group_id	account_name	budgeting	engine	ad_type	fob	theme	vendor_class	product_class_1	product_class_2	product_class_3	cta
