+++ test/data/shouldReadASingleLineAndReturnAChannel.csv
@@ -1,0 +1,2 @@
+channel	channel_account_id	channel_campaign_id	channel_adgroup_id	channel_ad_id	product_group_id	account_name	budgeting	engine	ad_type	fob	theme	vendor_class	product_class_1	product_class_2	product_class_3	cta
+Google	2831944540	186307692				GOOGLE_WOMENS_ACCESSORIES_PLA	WOMENS ACCESSORIES	GOOGLE	PLA	CENTER CORE			WOMENS ACCESSORIES	BRANDS	BRANDS
